+++ front/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "elpatochat",
+  "name": "chat",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -14,17 +14,16 @@
   },
   "dependencies": {
     "@fontsource/poppins": "^5.2.7",
-    "@tailwindcss/postcss": "^4.1.13",
+    "@stylistic/eslint-plugin": "^5.4.0",
     "@twurple/api": "^7.4.0",
     "@twurple/auth": "^7.4.0",
     "@twurple/chat": "^7.4.0",
     "@uidotdev/usehooks": "^2.4.1",
-    "motion": "^12.23.21",
-    "postcss": "^8.5.6",
+    "motion": "^12.23.22",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-router": "^7.9.2",
     "styled-components": "^6.1.19",
-    "tailwindcss": "^4.1.13",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
@@ -32,7 +31,6 @@
     "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.36.0",
     "@iconify/json": "^2.2.387",
-    "@iconify/tailwind": "^1.2.0",
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
     "@typescript-eslint/eslint-plugin": "^8.44.1",
@@ -40,7 +38,7 @@
     "@vitejs/plugin-react-swc": "^4.1.0",
     "eslint": "^9.36.0",
     "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-refresh": "^0.4.21",
+    "eslint-plugin-react-refresh": "^0.4.22",
     "globals": "^16.4.0",
     "typescript": "^5.9.2",
     "vite": "^7.1.7"
