--- front/postcss.config.js
@@ -1,5 +1,0 @@
-export default {
-  plugins: {
-    '@tailwindcss/postcss': {},
-  },
-};
