+++ front/src/components/ChatVisualizerCore/ChatBubble/chatHeader/styles.ts
