+++ front/src/components/ChatVisualizerCore/ChatBubble/chatMsgContent/styles.ts
