+++ front/src/main.tsx
@@ -4,7 +4,6 @@ import '@fontsource/poppins/800.css';
 import '@fontsource/poppins';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
-import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <App />
