+++ front/src/pages/Landing/index.tsx
@@ -1,0 +1,4 @@
+export const Landing = () => 
+{
+  return <div>El Pato Chat</div>;
+};
