--- front/tailwind.config.js
@@ -1,14 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-import { addDynamicIconSelectors } from '@iconify/tailwind';
-
-export default {
-  content: [
-    './index.html',
-    './src/**/*.{js,ts,jsx,tsx}'
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [addDynamicIconSelectors()]
-};
-
